+++ manager/controllers/pipeline_controller.go
@@ -34,7 +34,6 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 )
 
 // PipelineReconciler reconciles a Pipeline object.
@@ -92,7 +91,6 @@ func (r *PipelineReconciler) Reconcile(ctx context.Context, req ctrl.Request) (c
 			},
 			Spec: step,
 		}
-		controllerutil.AddFinalizer(obj, stepFinalizer)
 		if err := r.Client.Create(ctx, obj); err != nil {
 			if apierr.IsAlreadyExists(err) {
 				old := &dfv1.Step{}