+++ runner/sidecar/source/jetstream/jetstream.go
@@ -59,7 +59,7 @@ func New(ctx context.Context, secretInterface corev1.SecretInterface, cluster, n
 
 func (j jsSource) Close() error {
 	logger.Info("closing jetstream source connection")
-	if j.conn.IsClosed() {
+	if !j.conn.IsClosed() {
 		j.conn.Close()
 	}
 	return nil